# Fix calculation of best ops/sec in benchmarks