@SiarheiFedartsou 2024-06-29
I would really like to test it together with https://github.com/Project-OSRM/osrm-backend/pull/6975, so I am taking responsibility and merging it without review 🙏🏻